[Update] Move sum and prod into aggregate.py module
